[modify] 修改伤害为 int 类型 [add] 增加 DMG modfifier 类型，修饰最终伤害，他与hitbox互动

+++ Assets/Scripts/Units/Enemys/Turtle.cs
@@ -34,7 +34,7 @@ public class Turtle : GuardableEnemy
     private float m_slidingSpeed;
 
     private Team m_originalTeam;
-    private int m_originalDamage;
+    private float m_originalDamage;
     private LayerMask m_originalLayer;
 
     private Weapon m_selfWeapon;
